feat: add sign in, sign up and profile pages
