+++ app/(app)/profile.tsx
@@ -1,8 +1,9 @@
 import { useGetProfileCurrentQuery } from "@/store/slices/apiSlice";
 import { useTranslation } from "react-i18next";
-import { StyleSheet } from "react-native";
+import { ScrollView, StyleSheet } from "react-native";
 import {
     Card,
+    IconButton,
     List,
     ProgressBar,
     SegmentedButtons,
@@ -12,12 +13,15 @@ import {
 import { useDispatch, useSelector } from "react-redux";
 
 import { supportedLanguages } from "@/store/i18n";
+import { removeToken } from "@/store/slices/authenticationSlice";
 import { selectThemeMode, toggleThemeMode } from "@/store/slices/themeSlice";
 import { formatDate } from "@/store/utils/utilities";
+import { useRouter } from "expo-router";
 
 export default function Profile() {
   const { t, i18n } = useTranslation();
   const dispatch = useDispatch();
+  const router = useRouter();
   const themeMode = useSelector(selectThemeMode);
 
   const { currentData, isSuccess } = useGetProfileCurrentQuery(undefined, {
@@ -30,6 +34,11 @@ export default function Profile() {
 
   const handleChangeLanguage = (language: string) => {
     i18n.changeLanguage(language);
+  };
+
+  const signOut = () => {
+    dispatch(removeToken());
+    router.navigate("/sign-in");
   };
 
   const styles = StyleSheet.create({
@@ -55,78 +64,86 @@ export default function Profile() {
   });
 
   return (
-    <Surface style={styles.container} elevation={4}>
-      <Card style={styles.card}>
-        <Card.Title
-          title={t("profile.title", { ns: "authentication" })}
-          titleStyle={styles.title}
-          titleVariant="displayMedium"
-        />
-        {isSuccess && currentData !== undefined ? (
-          <Card.Content>
-            <List.Section>
-              <List.Subheader>
-                {t("profile.userData", { ns: "authentication" })}
-              </List.Subheader>
-              <List.Item
-                title={t("username.title", { ns: "authentication" })}
-                description={currentData.data?.username}
-                left={(props) => <List.Icon {...props} icon="human" />}
-              />
-              <List.Item
-                title={t("email.title", { ns: "authentication" })}
-                description={currentData.data?.email}
-                left={(props) => <List.Icon {...props} icon="email" />}
-              />
-              <List.Item
-                title={t("profile.memberSince", { ns: "authentication" })}
-                description={formatDate(
-                  i18n.language,
-                  currentData.data?.createdAt
-                )}
-                left={(props) => <List.Icon {...props} icon="calendar" />}
-              />
-            </List.Section>
-            <List.Section>
-              <List.Subheader>
-                {t("profile.settings", { ns: "authentication" })}
-              </List.Subheader>
-              <List.Item
-                title={t("theme.title")}
-                description={t(`theme.${themeMode}`)}
-                right={(props) => (
-                  <Switch
-                    {...props}
-                    value={themeMode === "light"}
-                    onValueChange={onToggleTheme}
-                  />
-                )}
-              />
-              <List.Item
-                title={""}
-                left={(props) => <List.Icon {...props} icon="translate" />}
-                right={(props) => (
-                  <SegmentedButtons
-                    value={i18n.language}
-                    onValueChange={handleChangeLanguage}
-                    buttons={supportedLanguages.map((language) => ({
-                      value: language,
-                      label: language.toUpperCase(),
-                    }))}
-                  />
-                )}
-              />
-            </List.Section>
-            <List.Section>
-              <List.Subheader>
-                {t("profile.actions", { ns: "authentication" })}
-              </List.Subheader>
-            </List.Section>
-          </Card.Content>
-        ) : (
-          <ProgressBar indeterminate={true} style={styles.loadingBar} />
-        )}
-      </Card>
-    </Surface>
+    <ScrollView>
+      <Surface style={styles.container} elevation={4}>
+        <Card style={styles.card}>
+          <Card.Title
+            title={t("profile.title", { ns: "authentication" })}
+            titleStyle={styles.title}
+            titleVariant="displayMedium"
+          />
+          {isSuccess && currentData !== undefined ? (
+            <Card.Content>
+              <List.Section>
+                <List.Subheader>
+                  {t("profile.userData", { ns: "authentication" })}
+                </List.Subheader>
+                <List.Item
+                  title={t("username.title", { ns: "authentication" })}
+                  description={currentData.data?.username}
+                  left={(props) => <List.Icon {...props} icon="human" />}
+                />
+                <List.Item
+                  title={t("email.title", { ns: "authentication" })}
+                  description={currentData.data?.email}
+                  left={(props) => <List.Icon {...props} icon="email" />}
+                />
+                <List.Item
+                  title={t("profile.memberSince", { ns: "authentication" })}
+                  description={formatDate(
+                    i18n.language,
+                    currentData.data?.createdAt
+                  )}
+                  left={(props) => <List.Icon {...props} icon="calendar" />}
+                />
+              </List.Section>
+              <List.Section>
+                <List.Subheader>
+                  {t("profile.settings", { ns: "authentication" })}
+                </List.Subheader>
+                <List.Item
+                  title={t("theme.title")}
+                  description={t(`theme.${themeMode}`)}
+                  right={(props) => (
+                    <Switch
+                      {...props}
+                      value={themeMode === "light"}
+                      onValueChange={onToggleTheme}
+                    />
+                  )}
+                />
+                <List.Item
+                  title={""}
+                  left={(props) => <List.Icon {...props} icon="translate" />}
+                  right={(props) => (
+                    <SegmentedButtons
+                      value={i18n.language}
+                      onValueChange={handleChangeLanguage}
+                      buttons={supportedLanguages.map((language) => ({
+                        value: language,
+                        label: language.toUpperCase(),
+                      }))}
+                    />
+                  )}
+                />
+              </List.Section>
+              <List.Section>
+                <List.Subheader>
+                  {t("profile.actions", { ns: "authentication" })}
+                </List.Subheader>
+                <List.Item
+                  title={t("signout", { ns: "authentication" })}
+                  right={(props) => (
+                    <IconButton {...props} icon="logout" onPress={signOut} />
+                  )}
+                />
+              </List.Section>
+            </Card.Content>
+          ) : (
+            <ProgressBar indeterminate={true} style={styles.loadingBar} />
+          )}
+        </Card>
+      </Surface>
+    </ScrollView>
   );
 }
