+++ store/locales/en/main.ts
@@ -1,7 +1,9 @@
 import authentication from "./authentication.json";
 import common from "./common.json";
+import error from "./error.json";
 
 export default {
   common: common,
   authentication: authentication,
+  error: error,
 };
