+++ store/slices/baseApiSlice.ts
@@ -28,7 +28,10 @@ const baseQueryWithAuthorizationHandling: BaseQueryFn<
   FetchBaseQueryError
 > = async (args, api, extraOptions) => {
   let result = await baseQuery(args, api, extraOptions);
-  if (result.error && result.error.status === 401) {
+  if (
+    result.error &&
+    (result.error.status === 401 || result.error.status === 403)
+  ) {
     api.dispatch(removeToken());
   }
   return result;
